Update .gitignore to ignore help tags
Without this, it causes some pretty interesting issues for anyone who uses the repo. Basically, they either diverge from the upstream/original history by committing the tags, or they are forced to constantly have them as untracked content. Other vim plugins tend to ignore them so that they don't have to be included in the repo.