Add test before remapping plugin's maps
The plugin maps the some keys to run the commands of the plugin
automatically by themselves.
Until now, those commands were always mapped, regardless of the state of
the plugin and the mapping themselves.

This caused two problems:
* When the plugin was not enabled the mapping still existed.
* When the mapping was empty, an error massage was generated.

To fix both of those problems, I added two tests before running the
mappings: first, it checks that the plugins is indeed enabled, and then
it checks that the mapping is not empty. Only in case both of those
tests passes, the commands are mapped.